CLOUD-1520: Don't use proxy for all communications at all
Communications intra-pods don't need to be proxied. In fact
they must not. This patch prevents traffic being routed to
a proxy in case a java property like "http.proxyHost" is set
in the environment.

As a side effect of this change, clustering will not be
affected by env variables like "https_proxy".